+++ app/Http/Controllers/Settings/SubscriptionsController.php
@@ -118,7 +118,7 @@ class SubscriptionsController extends Controller
             return redirect()->route('settings.index');
         }
 
-        auth()->user()->account->subscription(auth()->user()->account->getSubscribedPlanName())->cancelNow();
+        auth()->user()->account->subscriptionCancel();
 
         return redirect()->route('settings.subscriptions.downgrade.success');
     }
@@ -134,40 +134,15 @@ class SubscriptionsController extends Controller
             return redirect()->route('settings.index');
         }
 
-        $stripeToken = $request->input('stripeToken');
+        $result = auth()->user()->account->subscribe($request->input('stripeToken'), $request->input('plan'));
 
-        $plan = InstanceHelper::getPlanInformationFromConfig($request->input('plan'));
-        $errorMessage = '';
-
-        try {
-            auth()->user()->account->newSubscription($plan['name'], $plan['id'])
-                        ->create($stripeToken, [
-                            'email' => auth()->user()->email,
-                        ]);
-
+        if ($result === true) {
             return redirect()->route('settings.subscriptions.upgrade.success');
-        } catch (\Stripe\Error\Card $e) {
-            // Since it's a decline, \Stripe\Error\Card will be caught
-            $body = $e->getJsonBody();
-            $err = $body['error'];
-            $errorMessage = trans('settings.stripe_error_card', ['message' => $err['message']]);
-        } catch (\Stripe\Error\RateLimit $e) {
-            // Too many requests made to the API too quickly
-            $errorMessage = trans('settings.stripe_error_rate_limit');
-        } catch (\Stripe\Error\Authentication $e) {
-            // Authentication with Stripe's API failed
-            // (maybe you changed API keys recently)
-            $errorMessage = trans('settings.stripe_error_authentication');
-        } catch (\Stripe\Error\ApiConnection $e) {
-            // Network communication with Stripe failed
-            $errorMessage = trans('settings.stripe_error_api_connection_error');
-        } catch (\Stripe\Error\Base $e) {
-            $errorMessage = $e->getMessage();
         }
 
         return back()
             ->withInput()
-            ->withErrors($errorMessage);
+            ->withErrors($result);
     }
 
     /**